requirements.txt
flask
openai
duckduckgo-search
sentence-transformers
faiss-cpu
pypdf
Pillow
python-dotenv
requests
